% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizhours.R
\name{bizhours}
\alias{bizhours}
\title{bizhours}
\usage{
bizhours(from_dates, to_dates, start_time, end_time, timezone = "",
  calendar = "")
}
\arguments{
\item{from_dates}{A vector of POSIXct objects that are date-times from when to start
calculating business hours.}

\item{to_dates}{A vector of POSIXct objects that are date-times till when to
calculate business hours.}

\item{start_time}{Start of work shift. A string in format %H:%M that shows the time -
in 24 hour format - of the day when the business hours start.}

\item{end_time}{End of work shift. A string in format %H:%M that shows the time -
in 24 hour format- of the day when the business hours end.}

\item{timezone}{A string that represents the timezone. Default: "" - local timezone}

\item{calendar}{A string that represents the name of the \code{bizdays} calendar to use.
Default: "" - uses \code{Brazil/ANBIMA} calendar of package \code{bizdays} with
workweek from Monday through Friday.}
}
\value{
A numeric vector that shows the difference in business hours between
\code{from_dates} and \code{to_dates}
}
\description{
Calculate the business hours between two datetime vectors
}
\details{
Calculates the difference between two vectors of date times in terms of
business hours, where business hours is determined using the start_time and end_time
parameters. The function can be used seamlessly for night-shift calculations, eg. for 
a 9pm to 6am shift, \code{start_time} is 21:00 and \code{end_time} is 06:00
}
\author{
Sujeet G Pillai
}
\keyword{bizhours,}
\keyword{business}
\keyword{business,}
\keyword{calculate,}
\keyword{difference}
\keyword{difftime,}
\keyword{hours,}
\keyword{time,}
